refactor(engines): Update LoongArch64 engine and remove ABI1.0 support
