+++ src/engines/all.ts
@@ -4,4 +4,4 @@ import './win-arm64'
 import './xgo'
 import './freebsd'
 import './win7'
-import './loongarch-abi1.0'
+import './loongarch64'
